DE_PROJECT/requirements.txt
pymysql
psycopg2-binary
posthog==2.4.0
apache-airflow-providers-apache-spark